CLN: Renamed edit utilities to more sensical common
